update Generaldyne as measurement class
